requirements-test.txt
-r requirements.txt
coverage
mock
nose
django-nose
flake8
tox

